Lab-05_Starter/code/requirements.txt
Django>=5.1,<5.2
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
django-silk==5.5.0
